hw_4_7 : refactoring after review

+++ src/lesson4/hw/HW_4_7.java
@@ -18,6 +18,6 @@ public class HW_4_7 {
         for (int i = 0; i < array.length; i++) {
             sum += array[i];
         }
-        System.out.println(sum / array.length);
+        System.out.println((double) sum / array.length);
     }
 }
